Added all coloured blocks ((Sync Fix))
Now you can select the lock feature in the inspector, bump into the lock
and when you have the right key, it will open. Making 2 or more lock
sprites in an empty gameobject, makes you able to making a wall/cube of
locks, destroying them all when bumped into with the right keys.